Bp 3481 update test release v4.2.1

+++ upgrade/upgrade-4.2.1.php
@@ -41,20 +41,27 @@ function upgrade_module_4_2_1($object)
 
     Db::getInstance()->execute($createTableQuery);
 
-    // Move data from old table to new table
-    $moveDataQuery = 'INSERT INTO `' . _DB_PREFIX_ . 'bk_buckaroo_fee` (reference, id_cart, buckaroo_fee_tax_incl, buckaroo_fee_tax_excl, currency, created_at)
-    SELECT reference, id_cart, buckaroo_fee, buckaroo_fee, currency, created_at
-    FROM `' . _DB_PREFIX_ . 'buckaroo_fee`';
+    // Check if the old table exists
+    $tableExists = Db::getInstance()->executeS('SHOW TABLES LIKE "' . _DB_PREFIX_ . 'buckaroo_fee"');
+    if ($tableExists) {
+        // Move data from old table to new table
+        $moveDataQuery = 'INSERT INTO `' . _DB_PREFIX_ . 'bk_buckaroo_fee` (reference, id_cart, buckaroo_fee_tax_incl, buckaroo_fee_tax_excl, currency, created_at)
+        SELECT reference, id_cart, buckaroo_fee, buckaroo_fee, currency, created_at
+        FROM `' . _DB_PREFIX_ . 'buckaroo_fee`';
 
-    Db::getInstance()->execute($moveDataQuery);
+        Db::getInstance()->execute($moveDataQuery);
 
-    // Delete the old table
-    $deleteOldTableQuery = 'DROP TABLE IF EXISTS `' . _DB_PREFIX_ . 'buckaroo_fee`';
-    Db::getInstance()->execute($deleteOldTableQuery);
+        // Delete the old table
+        $deleteOldTableQuery = 'DROP TABLE IF EXISTS `' . _DB_PREFIX_ . 'buckaroo_fee`';
+        Db::getInstance()->execute($deleteOldTableQuery);
+    }
 
-    // Example of additional existing operations
-    Db::getInstance()->execute('ALTER TABLE `' . _DB_PREFIX_ . 'bk_giftcards` 
-    ADD is_custom INT(11) DEFAULT 0 NOT NULL;');
+    // Check if the column 'is_custom' already exists in 'bk_giftcards'
+    $columnExists = Db::getInstance()->executeS('SHOW COLUMNS FROM `' . _DB_PREFIX_ . 'bk_giftcards` LIKE "is_custom"');
+    if (empty($columnExists)) {
+        Db::getInstance()->execute('ALTER TABLE `' . _DB_PREFIX_ . 'bk_giftcards` 
+        ADD is_custom INT(11) DEFAULT 0 NOT NULL;');
+    }
 
     Db::getInstance()->execute('DELETE FROM ' . _DB_PREFIX_ . 'bk_payment_methods WHERE name = "tinka"');
